made Relation.size() return long instead of int to accommodate big BDDs

+++ runtime/src/jedd/internal/Backend.java
@@ -64,7 +64,7 @@ public abstract class Backend {
     abstract int numNodes( RelationInstance r );
     abstract int numPaths( RelationInstance r );
 
-    abstract int satCount( RelationInstance r, int vars );
+    abstract long satCount( RelationInstance r, int vars );
 
     abstract void gbc();
